step3 : Main 클래스에서 컨트롤러 분리
